requirements.txt
django==5.2.12
djangorestframework==3.16.1
django-redis==6.0.0
django-environ==0.12.0
django-tinymce==5.0.0
django-cors-headers==4.9.0

pyotp==2.9.0
qrcode==8.2
pillow==12.0.0

uvicorn==0.38.0

django-axes==8.0.0

channels==4.3.0
channels-redis==4.3.0

whitenoise==6.10.0

psycopg2==2.9.11


celery==5.6.0
django-celery-results==2.6.0
django-celery-beat==2.8.0


djoser==2.3.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
djangorestframework-simplejwt==5.5.1

django-storages==1.14.6
google-cloud-storage==3.1.0

Faker==38.0.0

bleach==6.3.0
beautifulsoup4==4.14.3